requirements.txt
alembic==0.9.6
awesome-slugify==1.6.5
click==6.7
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
ForgeryPy3==0.3.1
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
python-dateutil==2.6.1
python-editor==1.0.3
regex==2018.1.10
six==1.11.0
SQLAlchemy==1.1.14
Unidecode==0.4.21
Werkzeug==0.12.2
